+++ cmd/duo/main.go
@@ -22,7 +22,7 @@ import (
 	"github.com/atfa/duo/internal/workspace"
 )
 
-const version = "v0.3.1"
+const version = "v0.3.2"
 
 func main() {
 	if len(os.Args) > 1 {